+++ src/main/scala/com/gradle/SbtCommonCustomUserDataPlugin.scala
@@ -3,7 +3,7 @@ package com.gradle
 import com.gradle.develocity.agent.sbt.DevelocityPlugin
 import com.gradle.develocity.agent.sbt.DevelocityPlugin.autoImport.{DevelocityConfiguration, develocityConfiguration}
 import com.gradle.internal.{CustomBuildScanEnhancements, Overrides}
-import sbt.{AutoPlugin, Keys, Logger, Plugins, Setting, ScopeFilter, inAnyProject}
+import sbt.{AutoPlugin, Keys, Logger, Plugins, Setting}
 
 object SbtCommonCustomUserDataPlugin extends AutoPlugin {
 
@@ -14,24 +14,21 @@ object SbtCommonCustomUserDataPlugin extends AutoPlugin {
 
   override lazy val buildSettings: Seq[Setting[_]] = Seq(
     develocityConfiguration := {
-      val allScalaVersions = Keys.crossScalaVersions.all(ScopeFilter(inAnyProject)).value.flatten.distinct.sorted
       applyCCUD(
         Keys.sLog.value,
-        develocityConfiguration.value,
-        allScalaVersions
+        develocityConfiguration.value
       )
     }
   )
 
   private def applyCCUD(
       logger: Logger,
-      currentConfiguration: DevelocityConfiguration,
-      scalaVersions: Seq[String]
+      currentConfiguration: DevelocityConfiguration
   ): DevelocityConfiguration = {
     implicit val env: Env = Env.SystemEnvironment
     val transformers = Seq(
       Overrides.lift(_.server, _.withServer(_)),
-      CustomBuildScanEnhancements.transformer(scalaVersions, logger)
+      CustomBuildScanEnhancements.transformer(logger)
     )
 
     transformers